Consolidate dataset lineage modules
