+++ docs/Plan.md
@@ -542,7 +542,7 @@ Combine 1-4 and the *effective* context limit becomes hardware bandwidth, not mo
 50. **Parameter-efficient adaptation**: Explore low-rank fine-tuning across tasks; success is matching baseline accuracy with ≤10% extra parameters.
 51. **Context summarization memory**: Store compressed summaries for distant tokens and re-expand them on demand; success is >95% retrieval accuracy at 100× token length. *Implemented in `src/context_summary_memory.py` with tests.*
 51a. **Multi-modal summarization memory**: Compress image and audio features into short summaries stored with text embeddings; retrieval using fused summaries must reach ≥90% accuracy. *Implemented in `src/multimodal_summary_memory.py` with tests.*
-52. **Dataset lineage manager**: Automatically track dataset versions and transformations, enabling reproducible training pipelines. *Implemented in `src/dataset_lineage_manager.py`.*
+52. **Dataset lineage manager**: Automatically track dataset versions and transformations, enabling reproducible training pipelines. *Implemented in `src/dataset_lineage.py`.*
     Use `DataProvenanceLedger` to append a signed hash of each lineage record. Run `scripts/check_provenance.py <root>` to verify the ledger.
 52a. **Zero-trust memory server**: `ZeroTrustMemoryServer` validates signed access tokens against a `BlockchainProvenanceLedger` before serving requests. Unauthorized clients are rejected.
 52b. **Dataset watermarking**: `dataset_watermarker.py` embeds and detects watermarks in text, images and audio. Lineage records now include watermark IDs.
