+++ scripts/check_blockchain_provenance.py
@@ -1,7 +1,7 @@
 import argparse
 import json
 from pathlib import Path
-from asi.dataset_lineage_manager import DatasetLineageManager
+from asi.dataset_lineage import DatasetLineageManager
 from asi.blockchain_provenance_ledger import BlockchainProvenanceLedger
