+++ src/__init__.py
@@ -186,7 +186,7 @@ from .telemetry_aggregator import TelemetryAggregator
 from .cognitive_load_monitor import CognitiveLoadMonitor
 from .license_inspector import LicenseInspector
 from .dataset_versioner import DatasetVersioner
-from .dataset_lineage_manager import DatasetLineageManager
+from .dataset_lineage import DatasetLineageManager
 from .dataset_lineage_dashboard import DatasetLineageDashboard
 from .dataset_anonymizer import DatasetAnonymizer
 from .ner_anonymizer import NERAnonymizer
