+++ tests/test_dataset_lineage_dashboard.py
@@ -31,7 +31,7 @@ def _load(name: str, path: str):
     loader.exec_module(mod)
     return mod
 
-DatasetLineageManager = _load('src.dataset_lineage_manager', 'src/dataset_lineage_manager.py').DatasetLineageManager
+DatasetLineageManager = _load('src.dataset_lineage', 'src/dataset_lineage.py').DatasetLineageManager
 DatasetLineageDashboard = _load('src.dataset_lineage_dashboard', 'src/dataset_lineage_dashboard.py').DatasetLineageDashboard
