+++ tests/test_dataset_lineage_manager.py
@@ -13,11 +13,11 @@ src_pkg = types.ModuleType('src')
 sys.modules['src'] = src_pkg
 src_pkg.__path__ = ['src']
 src_pkg.__spec__ = importlib.machinery.ModuleSpec('src', None, is_package=True)
-loader = importlib.machinery.SourceFileLoader('src.dataset_lineage_manager', 'src/dataset_lineage_manager.py')
+loader = importlib.machinery.SourceFileLoader('src.dataset_lineage', 'src/dataset_lineage.py')
 spec = importlib.util.spec_from_loader(loader.name, loader)
 dlm = importlib.util.module_from_spec(spec)
 dlm.__package__ = 'src'
-sys.modules['src.dataset_lineage_manager'] = dlm
+sys.modules['src.dataset_lineage'] = dlm
 loader.exec_module(dlm)
 DatasetLineageManager = dlm.DatasetLineageManager
